a lot of changes, some large, some small.
The three most important things, and what caused a lot of files to change is the
ability to receive XML forecasts now, bringing up the XML output to the current
standard, and receiving both text and XML reports for turn 0.

The default is now to send both versions of the report in order to support
turn viewers that only deal with XML. To that end there's a new option added:
TXTReport. By default it's on, but can be turned off.