additional note: the -due command was implemented which allows a GM to send out an email to all the players that have not yet sent in orders.
Hardcoded message tells user that the turn will run in less than 24
hours.  Typically this command will be run from a cron job to alert
users. The GM will also receive a copy of the email, but it won't say
who the users are - just use it as a notice that not all orders are
in. Use the galaxyng -lastorders to see exactly which players got the
email.