implemented PaxGalactica
This means that you can now put a Peace <number of turns> directive in
the .glx file and the engine will ensure that no one can declare war
on another race until that many turns has expired.

In theory this gives everyone the chance to build up a fleet before
sparring starts.

I think that someone could end up with a fleet at their doorstep just
as the PaxGalactica ends.

The engine will *not* make everyone enemies at the end of the imposed
peace - it just allows the players to set their status war then.