added code to truncate player names if they get too long.
This version of the mapping software drops the legend below the map
proper. This was done for a couple of reasons: 1) it means that the
map itself can be larger and still fit on a 1024x768 browser page
without horizontal scrolling; 2) the maps that don't show player names
can now just drop that section off without having a lot of blank
space.

This is still a fairly memory intensive process - expect another
release that may fix that problem.