New feature added: -immediate
This feature will run a check to see if one of two conditions has been
met: a) all players have submitted orders, or b) the GM specified
number of hours has elapsed since the last tick was run.

In order to accomplish this a command has been added (-immediate) and
a key has been added to the .galaxyngrc file.

The key is the tick interval in hours and has the form:

<gamename>.tick  <tick interval in hours>

So, for example, it would look like:

SpaceInvaders.tick  48
Morat1.tick 72